Changes to folder structure, pyinvoke file, and some method names
