+++ MANIFEST.in
@@ -1,2 +1,2 @@
-include policy_sentry/shared/links.yml
+include policy_sentry/shared/data/links.yml
 recursive-include policy_sentry/shared *.txt *.html *.yml
