+++ docs/appendix/policy_sentry-as-a-python-package.rst
@@ -2,3 +2,5 @@ Policy Sentry as a Python Package
 ########################################
 
 **TODO: Instructions for using it as a python package so you can query the database.**
+
+**TODO: Change the initialization function so that there is an option to use a database file bundled with the package for quick import as a library**
