+++ policy_sentry/command/write_policy_dir.py
@@ -7,7 +7,7 @@ import glob
 import click
 from policy_sentry.shared.database import connect_db
 from policy_sentry.command.write_policy import write_policy_with_actions, write_policy_with_access_levels
-from policy_sentry.shared.file import read_yaml_file, write_json_file, check_valid_file_path
+from policy_sentry.util.file import read_yaml_file, write_json_file, check_valid_file_path
 from policy_sentry.shared.constants import DATABASE_FILE_PATH
