+++ policy_sentry/querying/conditions.py
@@ -1,0 +1,31 @@
+"""
+Methods that execute specific queries against the SQLite database for the CONDITIONS table.
+This supports the policy_sentry query functionality
+"""
+from sqlalchemy import and_
+from policy_sentry.shared.database import ConditionTable
+
+
+# Per service
+def get_condition_keys_for_service(db_session, service):
+    """Get a list of available conditions per AWS service"""
+    results = []
+    rows = db_session.query(ConditionTable.condition_key_name, ConditionTable.condition_value_type,
+                            ConditionTable.description).filter(ConditionTable.service.like(service))
+    for row in rows:
+        results.append(str(row.condition_key_name))
+    return results
+
+
+# Per condition key name
+def get_condition_key_details(db_session, service, condition_key_name):
+    """Get details about a specific condition key in JSON format"""
+    rows = db_session.query(ConditionTable.condition_key_name, ConditionTable.condition_value_type,
+                            ConditionTable.description).filter(and_(ConditionTable.condition_key_name.like(condition_key_name), ConditionTable.service.like(service)))
+    result = rows.first()
+    output = {
+        'name': result.condition_key_name,
+        'description': result.description,
+        'condition_value_type': result.condition_value_type
+    }
+    return output
