+++ policy_sentry/util/conditions.py
