+++ policy_sentry/writing/minimize.py
