+++ test/test_access_level_overrides.py
@@ -1,5 +1,6 @@
 import unittest
-from policy_sentry.shared.config import get_action_access_level_overrides_from_yml, determine_access_level_override
+from policy_sentry.util.access_levels import determine_access_level_override
+from policy_sentry.configuration.access_level_overrides import get_action_access_level_overrides_from_yml
 
 
 class AccessLevelOverride(unittest.TestCase):
