+++ test/test_arn_action_group.py
@@ -1,6 +1,6 @@
 import unittest
 from policy_sentry.shared.database import connect_db
-from policy_sentry.shared.policy import ArnActionGroup
+from policy_sentry.writing.policy import ArnActionGroup
 from policy_sentry.shared.constants import DATABASE_FILE_PATH
 
 db_session = connect_db(DATABASE_FILE_PATH)
