+++ test/test_expand_wildcard_actions.py
@@ -1,4 +1,4 @@
-from policy_sentry.shared.analyze import determine_actions_to_expand
+from policy_sentry.analysis.analyze import determine_actions_to_expand
 import unittest
