+++ test/test_template.py
@@ -1,5 +1,5 @@
 import unittest
-from policy_sentry.shared.template import create_actions_template, create_crud_template
+from policy_sentry.writing.template import create_actions_template, create_crud_template
 
 
 class TemplateTestCase(unittest.TestCase):
