+++ test/test_write_policy.py
@@ -1,9 +1,9 @@
 import unittest
 import json
 from policy_sentry.shared.database import connect_db
-from policy_sentry.shared.policy import ArnActionGroup
+from policy_sentry.writing.policy import ArnActionGroup
 from policy_sentry.command.write_policy import print_policy
-from policy_sentry.shared.actions import get_dependent_actions
+from policy_sentry.querying.actions import get_dependent_actions
 from policy_sentry.shared.constants import DATABASE_FILE_PATH
 
 db_session = connect_db(DATABASE_FILE_PATH)
